Split api routes in different files
