+++ views/singup.handlebars
@@ -1,5 +1,5 @@
-<form class="login-form">
-  <div class="input-group">
+<form class="singup-form">
+  <div class="singup-group">
     <label for="username">Username</label>
     <input type="text" id="username" name="username" />
   </div>
@@ -7,13 +7,7 @@
     <label for="password">Password</label>
     <input type="password" id="password" name="password" />
   </div>
-  <button type="submit">Login</button>
+  <button type="submit">Sing up!</button>
 </form>
-<div>
-  <a href="/singup">Sign up instead</a>
-</div>
-{{!-- <main>
-  {{{body}}}
-</main> --}}
 
 <script src="./js/login.js"></script>
